requirements.txt
Flask
torch
torchvision
transformers
opencv-python
Pillow
numpy
datasets
tensorflow
efficientnet